Removed big O note for summed area table
It requires more explanation, and I don't want to make this readme
longer.

It was saying O(N^2), but what is N? It was the largest side of the
rectangle in pixels. I guess you could say, the performance is O(N),
where N is number of pixels. But then have to stress out that we are
talking about the entire screen (which in many cases was more than 2.3M
pixels)